Added alert functionality using sweetalert2

+++ src/app/MyComponents/todos/todos.component.ts
@@ -1,5 +1,6 @@
 import { Todo } from './../../Todo';
 import { Component, EventEmitter, OnInit, Output } from '@angular/core';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-todos',
@@ -19,19 +20,54 @@ export class TodosComponent implements OnInit {
   }
   ngOnInit(): void {}
   deleteTodo(todo: Todo) {
-    console.log(todo);
-    const index = this.todos.indexOf(todo);
-    this.todos.splice(index, 1);
-    localStorage.setItem('todos', JSON.stringify(this.todos));
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire('Deleted!', todo.title+' has been deleted.', 'success');
+        const index = this.todos.indexOf(todo);
+        this.todos.splice(index, 1);
+        localStorage.setItem('todos', JSON.stringify(this.todos));
+      }
+    });
   }
   deleteAll() {
-    this.todos.splice(0, this.todos.length);
-    localStorage.setItem('todos', JSON.stringify(this.todos));
+    if (this.todos.length != 0) {
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete All!',
+      }).then((result) => {
+        if (result.isConfirmed) {
+          Swal.fire('Deleted!', 'All Todos have been deleted.', 'success');
+          this.todos.splice(0, this.todos.length);
+          localStorage.setItem('todos', JSON.stringify(this.todos));
+        }
+      });
+    }
+    else{
+      Swal.fire('List is Already Empty!!', '', 'error')
+    }
   }
   addTodo(todo: Todo) {
-    console.log(todo);
-    this.todos.push(todo);
-    localStorage.setItem('todos', JSON.stringify(this.todos));
+    if (todo.title != null || todo.desc != null) {
+      this.todos.push(todo);
+      localStorage.setItem('todos', JSON.stringify(this.todos));
+      Swal.fire(todo.title+' Added Successfully', '', 'success');
+    }
+    else{
+      Swal.fire('A field is left empty', '', 'error');
+    }
   }
   toggleTodo(todo: Todo) {
     const index = this.todos.indexOf(todo);
